refactor(scheduler): extract utility functions and types for clarity
