+++ src/qdash/workflow/engine/scheduler/cr_scheduler.py
@@ -40,15 +40,21 @@ Example:
 
 from __future__ import annotations
 
-import itertools
 import logging
-from collections import defaultdict
 from pathlib import Path
 from typing import TYPE_CHECKING, Any
 
-import networkx as nx
 import yaml
 from qdash.workflow.engine.backend.qubex_paths import get_qubex_paths
+from qdash.workflow.engine.scheduler.cr_utils import (
+    build_mux_conflict_map,
+    build_qubit_to_mux_map,
+    convert_to_parallel_groups,
+    extract_qubit_frequency,
+    group_cr_pairs_by_conflict,
+    infer_direction_from_design,
+    split_fast_slow_pairs,
+)
 
 if TYPE_CHECKING:
     from qdash.common.topology_config import TopologyDefinition
@@ -262,19 +268,6 @@ class CRScheduler:
 
         return self._wiring_config
 
-    @staticmethod
-    def _extract_qubit_frequency(qubits: dict[str, Any]) -> dict[str, float]:
-        """Extract qubit frequencies from qubit models.
-
-        Works with both QubitModel (embedded) and QubitDocument (individual collection).
-        Both have .data attribute with calibration data.
-        """
-        return {
-            qid: qubit.data["qubit_frequency"]["value"]
-            for qid, qubit in qubits.items()
-            if qubit.data and "qubit_frequency" in qubit.data
-        }
-
     def _get_two_qubit_pair_list(self) -> list[str]:
         """Extract all two-qubit coupling IDs from CouplingDocument collection."""
         coupling_ids = self._load_coupling_ids()
@@ -283,240 +276,6 @@ class CRScheduler:
             for coupling_id in coupling_ids
             if "-" in coupling_id and len(coupling_id.split("-")) == 2
         ]
-
-    @staticmethod
-    def _build_mux_conflict_map(yaml_mux_list: list[dict[str, Any]]) -> dict[int, set[int]]:
-        """Build conflict map for MUX resources.
-
-        MUXes conflict if they share the same readout or control module.
-        """
-        module_to_muxes_readout: dict[str, set[int]] = defaultdict(set)
-        module_to_muxes_ctrl: dict[str, set[int]] = defaultdict(set)
-
-        # Group MUXes by readout and control modules
-        for mux_entry in yaml_mux_list:
-            mux_id = mux_entry["mux"]
-
-            # Readout module conflicts
-            read_out = mux_entry.get("read_out")
-            if read_out:
-                readout_module = read_out.split("-")[0]
-                module_to_muxes_readout[readout_module].add(mux_id)
-
-            # Control module conflicts
-            for ctrl in mux_entry.get("ctrl", []):
-                ctrl_module = ctrl.split("-")[0]
-                module_to_muxes_ctrl[ctrl_module].add(mux_id)
-
-        def create_conflict_map(module_to_muxes: dict[str, set[int]]) -> dict[int, set[int]]:
-            """Create bidirectional conflict map from module groupings."""
-            mux_conflict: dict[int, set[int]] = defaultdict(set)
-            for muxes in module_to_muxes.values():
-                for mux_a, mux_b in itertools.combinations(muxes, 2):
-                    mux_conflict[mux_a].add(mux_b)
-                    mux_conflict[mux_b].add(mux_a)
-            return mux_conflict
-
-        # Merge readout and control conflicts
-        conflict_map = create_conflict_map(module_to_muxes_readout)
-        ctrl_conflict_map = create_conflict_map(module_to_muxes_ctrl)
-
-        for mux_id, conflicts in ctrl_conflict_map.items():
-            conflict_map[mux_id].update(conflicts)
-
-        return dict(conflict_map)
-
-    @staticmethod
-    def _build_qubit_to_mux_map(yaml_mux_list: list[dict[str, Any]]) -> dict[str, int]:
-        """Build mapping from qubit ID to MUX ID.
-
-        Each MUX controls 4 qubits: MUX_N controls qubits [4N, 4N+1, 4N+2, 4N+3].
-        """
-        qid_to_mux = {}
-        for entry in yaml_mux_list:
-            mux_id = entry["mux"]
-            for offset in range(4):
-                qid_to_mux[str(mux_id * 4 + offset)] = mux_id
-        return qid_to_mux
-
-    @staticmethod
-    def _qid_to_coords(qid: int, grid_size: int) -> tuple[int, int]:
-        """Convert qubit ID to (row, col) coordinates in the square lattice.
-
-        Args:
-            qid: Qubit ID (0-indexed)
-            grid_size: Grid dimension (8 for 64-qubit, 12 for 144-qubit)
-
-        Returns:
-            (row, col) tuple representing position in the grid
-
-        Example:
-            For 64-qubit chip (8x8 grid):
-            - qid=0 → (0, 0) [MUX 0, position TL]
-            - qid=1 → (0, 1) [MUX 0, position TR]
-            - qid=2 → (1, 0) [MUX 0, position BL]
-            - qid=16 → (2, 0) [MUX 4, position TL]
-        """
-        # Which MUX does this qubit belong to?
-        mux_id = qid // 4
-
-        # Position within the MUX (0=TL, 1=TR, 2=BL, 3=BR)
-        pos_in_mux = qid % 4
-
-        # MUX grid dimension (N/2 × N/2)
-        mux_grid_size = grid_size // 2
-
-        # MUX position in MUX grid
-        mux_row = mux_id // mux_grid_size
-        mux_col = mux_id % mux_grid_size
-
-        # Position within MUX (2×2 sub-grid)
-        local_row = pos_in_mux // 2  # 0 (top) or 1 (bottom)
-        local_col = pos_in_mux % 2  # 0 (left) or 1 (right)
-
-        # Combine to get global position
-        row = mux_row * 2 + local_row
-        col = mux_col * 2 + local_col
-
-        return (row, col)
-
-    @staticmethod
-    def _infer_direction_from_design(qid1: str, qid2: str, grid_size: int = 8) -> bool:
-        """Infer CR gate direction from design-based frequency pattern.
-
-        The chip follows a checkerboard frequency pattern where frequency is determined by
-        coordinate parity. This allows inferring CR direction without actual frequency measurements.
-
-        Design pattern (from docs/architecture/square-lattice-topology.md):
-        - Low frequency (~8000 MHz): (row + col) % 2 == 0
-        - High frequency (~9000 MHz): (row + col) % 2 == 1
-
-        CR gate constraint: f_control < f_target
-
-        Args:
-            qid1: First qubit ID
-            qid2: Second qubit ID
-            grid_size: Grid dimension (8 for 64-qubit, 12 for 144-qubit, default: 8)
-
-        Returns:
-            True if qid1 should be control (qid1 has lower frequency by design),
-            False otherwise
-
-        Example:
-            For 64-qubit chip:
-            - qid1=0 → (0,0) → sum=0 (even) → low freq
-            - qid2=1 → (0,1) → sum=1 (odd) → high freq
-            - Result: True (0 is control, 1 is target)
-        """
-        r1, c1 = CRScheduler._qid_to_coords(int(qid1), grid_size)
-        r2, c2 = CRScheduler._qid_to_coords(int(qid2), grid_size)
-
-        # Checkerboard pattern: (row + col) % 2 determines frequency group
-        # Even sum → low frequency, Odd sum → high frequency
-        parity1 = (r1 + c1) % 2
-        parity2 = (r2 + c2) % 2
-
-        # CR constraint: control has lower frequency
-        # parity=0 → low freq, parity=1 → high freq
-        return parity1 < parity2
-
-    @staticmethod
-    def _group_cr_pairs_by_conflict(
-        cr_pairs: list[str],
-        qid_to_mux: dict[str, int],
-        mux_conflict_map: dict[int, set[int]],
-        max_parallel_ops: int | None = None,
-        coloring_strategy: str = "largest_first",
-    ) -> list[list[str]]:
-        """Group CR pairs into parallel execution steps using greedy graph coloring.
-
-        Args:
-            cr_pairs: List of CR pair strings (e.g., ["0-1", "2-3"])
-            qid_to_mux: Mapping from qubit ID to MUX ID
-            mux_conflict_map: MUX conflict relationships
-            max_parallel_ops: Maximum parallel operations per group
-            coloring_strategy: NetworkX graph coloring strategy. Options:
-                - "largest_first": Largest degree first (default, good general performance)
-                - "smallest_last": Smallest degree last (often better quality)
-                - "random_sequential": Random order (non-deterministic)
-                - "connected_sequential_bfs": BFS ordering
-                - "connected_sequential_dfs": DFS ordering
-                - "saturation_largest_first": DSATUR algorithm (often optimal)
-
-        Returns:
-            List of groups where each group contains CR pairs that can run in parallel
-        """
-        # Build conflict graph
-        conflict_graph = nx.Graph()
-        conflict_graph.add_nodes_from(cr_pairs)
-
-        for pair_a, pair_b in itertools.combinations(cr_pairs, 2):
-            q1a, q2a = pair_a.split("-")
-            q1b, q2b = pair_b.split("-")
-
-            # Conflict 1: Shared qubits
-            if {q1a, q2a} & {q1b, q2b}:
-                conflict_graph.add_edge(pair_a, pair_b)
-                continue
-
-            # Conflict 2: Same MUX usage
-            mux_a1, mux_a2 = qid_to_mux[q1a], qid_to_mux[q2a]
-            mux_b1, mux_b2 = qid_to_mux[q1b], qid_to_mux[q2b]
-
-            if mux_a1 in (mux_b1, mux_b2) or mux_a2 in (mux_b1, mux_b2):
-                conflict_graph.add_edge(pair_a, pair_b)
-                continue
-
-            # Conflict 3: MUX resource conflicts
-            conflict_muxes = mux_conflict_map.get(mux_a1, set()) | mux_conflict_map.get(
-                mux_a2, set()
-            )
-            if mux_b1 in conflict_muxes or mux_b2 in conflict_muxes:
-                conflict_graph.add_edge(pair_a, pair_b)
-
-        # Greedy graph coloring
-        coloring = nx.coloring.greedy_color(conflict_graph, strategy=coloring_strategy)
-
-        # Group pairs by color
-        color_groups: dict[int, list[str]] = defaultdict(list)
-        for pair, color in coloring.items():
-            color_groups[color].append(pair)
-
-        # Convert to sorted list of groups
-        groups = [color_groups[c] for c in sorted(color_groups)]
-
-        # Optionally split groups that exceed max parallel operations limit
-        if max_parallel_ops is not None:
-            split_groups = []
-            for group in groups:
-                for i in range(0, len(group), max_parallel_ops):
-                    chunk = group[i : i + max_parallel_ops]
-                    split_groups.append(chunk)
-            return split_groups
-
-        return groups
-
-    @staticmethod
-    def _split_fast_slow_pairs(
-        cr_pairs: list[str], qid_to_mux: dict[str, int]
-    ) -> tuple[list[str], list[str]]:
-        """Separate CR pairs into fast (intra-MUX) and slow (inter-MUX) categories."""
-        fast_pairs = [
-            p
-            for p in cr_pairs
-            if qid_to_mux.get(p.split("-")[0]) == qid_to_mux.get(p.split("-")[1])
-        ]
-        slow_pairs = [
-            p
-            for p in cr_pairs
-            if qid_to_mux.get(p.split("-")[0]) != qid_to_mux.get(p.split("-")[1])
-        ]
-        return fast_pairs, slow_pairs
-
-    @staticmethod
-    def _convert_to_parallel_groups(grouped: list[list[str]]) -> list[list[tuple[str, str]]]:
-        """Convert grouped CR pairs to parallel_groups format."""
-        return [[(pair.split("-")[0], pair.split("-")[1]) for pair in group] for group in grouped]
 
     def generate_with_plugins(
         self,
@@ -586,12 +345,12 @@ class CRScheduler:
         # Load chip metadata and qubit data from individual documents (scalable)
         chip = self._load_chip_data()
         qubit_models = self._load_qubit_models()
-        qubit_frequency = self._extract_qubit_frequency(qubit_models)
+        qubit_frequency = extract_qubit_frequency(qubit_models)
 
         # Load MUX configuration
         wiring_config = self._load_wiring_config()
-        mux_conflict_map = self._build_mux_conflict_map(wiring_config)
-        qid_to_mux = self._build_qubit_to_mux_map(wiring_config)
+        mux_conflict_map = build_mux_conflict_map(wiring_config)
+        qid_to_mux = build_qubit_to_mux_map(wiring_config)
 
         # Determine grid size
         grid_size = 12 if "144Q" in self.chip_id else 8
@@ -656,10 +415,10 @@ class CRScheduler:
         scheduler_metadata = scheduler.get_metadata()
 
         # Convert to parallel_groups format
-        parallel_groups = self._convert_to_parallel_groups(grouped)
+        parallel_groups = convert_to_parallel_groups(grouped)
 
         # Calculate fast/slow split for metadata
-        fast, slow = self._split_fast_slow_pairs(filtered_pairs, qid_to_mux)
+        fast, slow = split_fast_slow_pairs(filtered_pairs, qid_to_mux)
 
         logger.info(
             f"Generated schedule: {sum(len(g) for g in parallel_groups)} pairs in {len(parallel_groups)} groups"
@@ -757,7 +516,7 @@ class CRScheduler:
 
         # Load qubit data from individual documents (scalable)
         qubit_models = self._load_qubit_models()
-        qubit_frequency = self._extract_qubit_frequency(qubit_models)
+        qubit_frequency = extract_qubit_frequency(qubit_models)
 
         logger.debug(f"  Total qubits in chip: {len(qubit_models)}")
         logger.debug(f"  Qubits with frequency data: {len(qubit_frequency)}")
@@ -803,7 +562,7 @@ class CRScheduler:
                     for pair in all_pairs
                     if (qubits := pair.split("-"))
                     and len(qubits) == 2
-                    and not self._infer_direction_from_design(qubits[0], qubits[1], grid_size)
+                    and not infer_direction_from_design(qubits[0], qubits[1], grid_size)
                 ]
             else:
                 cr_pairs = [
@@ -811,7 +570,7 @@ class CRScheduler:
                     for pair in all_pairs
                     if (qubits := pair.split("-"))
                     and len(qubits) == 2
-                    and self._infer_direction_from_design(qubits[0], qubits[1], grid_size)
+                    and infer_direction_from_design(qubits[0], qubits[1], grid_size)
                 ]
         else:
             direction_method = "measured"
@@ -846,8 +605,8 @@ class CRScheduler:
 
         # Load MUX configuration
         wiring_config = self._load_wiring_config()
-        mux_conflict_map = self._build_mux_conflict_map(wiring_config)
-        qid_to_mux = self._build_qubit_to_mux_map(wiring_config)
+        mux_conflict_map = build_mux_conflict_map(wiring_config)
+        qid_to_mux = build_qubit_to_mux_map(wiring_config)
 
         # Filter out CR pairs that have qubits without MUX mappings
         pairs_before_mux_filter = len(cr_pairs)
@@ -864,15 +623,15 @@ class CRScheduler:
             raise ValueError(msg)
 
         # Group pairs: fast (intra-MUX) first, then slow (inter-MUX)
-        fast, slow = self._split_fast_slow_pairs(cr_pairs, qid_to_mux)
-        grouped = self._group_cr_pairs_by_conflict(
+        fast, slow = split_fast_slow_pairs(cr_pairs, qid_to_mux)
+        grouped = group_cr_pairs_by_conflict(
             fast, qid_to_mux, mux_conflict_map, max_parallel_ops, coloring_strategy
-        ) + self._group_cr_pairs_by_conflict(
+        ) + group_cr_pairs_by_conflict(
             slow, qid_to_mux, mux_conflict_map, max_parallel_ops, coloring_strategy
         )
 
         # Convert to parallel_groups format
-        parallel_groups = self._convert_to_parallel_groups(grouped)
+        parallel_groups = convert_to_parallel_groups(grouped)
 
         logger.info(
             f"Generated schedule: {sum(len(g) for g in parallel_groups)} pairs in {len(parallel_groups)} groups"
